+++ pysui/sui/sui_pgql/config/__init__.py
@@ -1,0 +1,6 @@
+#    Copyright Frank V. Castellucci
+#    SPDX-License-Identifier: Apache-2.0
+
+# -*- coding: utf-8 -*-
+
+"""Pysui Sui GraphQL Configuration Package."""